[Fix] determining which files need to be moved
