+++ modules/shared.py
